feat: log ingest is working
Refactor to cluster.local dnsDomain because OpenTelemetry Operator hard
codes the target allocator service name for cluster.local, should shoot
them an issue about it and fix it.
Refactor to loki exporter, which is deprecated, but the otlp ingest in
Loki simply refuses to pick labels up no matter how I configure it. It
doesn't even make sense reading it's source code.
Refactor to use custom plugin escape hatch in Fluent operator to set an
option they haven't yet added to the operator. Should shoot an issue and
PR for it later.